// src/shared/constants/enum.ts
export enum RoleType {
  ADMIN = 'admin',
  USER = 'user',
}

export enum DefaultIpColumn {
  CREATED_AT_IP = 'created_at_ip',
  UPDATED_AT_IP = 'updated_at_ip',
  DELETED_AT_IP = 'deleted_at_ip',
}

export enum DefaultStatus {
  ACTIVE = 'active',
  IN_ACTIVE = 'in_active',
}

export enum OtpType {
  SIGNUP = 'signup',
  LOGIN = 'login',
  FORGOT_PASSWORD = 'forgot_password',
}